Remove HtmxViewHandlerInterceptor
The HtmxResponseHandlerMethodReturnValueHandler alone is enough to cover the HtmxResponse return value from normal controller methods and exception handlers

Fixes #94
