+++ htmx-spring-boot/src/test/java/io/github/wimdeblauwe/htmx/spring/boot/mvc/HtmxResponseHandlerMethodReturnValueHandlerController.java
@@ -12,7 +12,7 @@ import java.util.TreeMap;
 
 @Controller
 @RequestMapping("/hvhi")
-public class HtmxViewHandlerInterceptorController {
+public class HtmxResponseHandlerMethodReturnValueHandlerController {
 
     @GetMapping("/hx-location-with-context-data")
     public HtmxResponse hxLocationWithContextData() {
